Moved class to library package and applied PR refactorings to MergeImages (e.g. more speaking names)
